requirements.txt
langgraph>=0.0.40
langchain>=0.1.0
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0
pydantic>=2.0.0
networkx>=3.0
numpy>=1.21.0
chromadb>=0.4.0
neo4j>=5.0.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
